Fixed/silenced a few implicit narrowing warnings in the 32-bit MSVC build.
Mostly converting off64_t/uint64_t to size_t.

+++ src/libromdata/Other/Wim.cpp
@@ -191,8 +191,8 @@ int WimPrivate::addFields_XML()
 		return -EIO;
 	}
 
-	char16_t *const xml_data = new char16_t[size/2];
-	size_t bytes_read = file->read(xml_data, size);
+	char16_t *const xml_data = new char16_t[static_cast<size_t>(size / 2)];
+	size_t bytes_read = file->read(xml_data, static_cast<size_t>(size));
 	if (bytes_read != size) {
 		delete[] xml_data;
 		return -EIO;
@@ -200,7 +200,7 @@ int WimPrivate::addFields_XML()
 
 	// the xml inside wims are utf-16 but tinyxml only supports utf-8
 	// this means we have to do some conversion
-	string utf8_xml = LibRpText::utf16le_to_utf8(xml_data, size/2);
+	string utf8_xml = LibRpText::utf16le_to_utf8(xml_data, static_cast<size_t>(size / 2));
 	delete[] xml_data;
 
 	XMLDocument document;
